feat: remove sizes field from product creation
- Removed the `sizes` input field from `frontend/admin.html` and its handling in `frontend/js/admin.js`.
- Removed the `sizes` column from `PhotoDB` model in `backend/src/models.py`.
- Removed the `sizes` field from the `Photo` Pydantic schema in `backend/src/schemas.py`.
- Re-generated the initial Alembic migration (`dd3948994de5_initial_migration.py`) to reflect the removal of the `sizes` column.
- Added safety checks `inspector.has_table(...)` in the migration script to prevent crashes when migrating against an existing database that already has these tables.
- Removed unused `import json` from `backend/src/routers/item.py`.
