--- backend/alembic/versions/99aad75c190d_initial_migration.py
@@ -1,71 +1,0 @@
-"""initial_migration
-
-Revision ID: 99aad75c190d
-Revises:
-Create Date: 2026-03-14 07:47:27.613836
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '99aad75c190d'
-down_revision: Union[str, Sequence[str], None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('order_items',
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('order_id', sa.String(), nullable=True),
-    sa.Column('photo_id', sa.String(), nullable=True),
-    sa.Column('quantity', sa.Integer(), nullable=True),
-    sa.Column('size', sa.String(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_order_items_id'), 'order_items', ['id'], unique=False)
-    op.create_table('orders',
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('email', sa.String(), nullable=True),
-    sa.Column('firstName', sa.String(), nullable=True),
-    sa.Column('lastName', sa.String(), nullable=True),
-    sa.Column('address', sa.String(), nullable=True),
-    sa.Column('phone', sa.String(), nullable=True),
-    sa.Column('deliveryMethod', sa.String(), nullable=True),
-    sa.Column('paymentMethod', sa.String(), nullable=True),
-    sa.Column('novaPoshta', sa.JSON(), nullable=True),
-    sa.Column('messenger', sa.String(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_orders_id'), 'orders', ['id'], unique=False)
-    op.create_table('photos',
-    sa.Column('id', sa.String(), nullable=False),
-    sa.Column('filename', sa.String(), nullable=True),
-    sa.Column('name', sa.String(), nullable=True),
-    sa.Column('item_description', sa.String(), nullable=True),
-    sa.Column('path', sa.String(), nullable=True),
-    sa.Column('price', sa.Float(), nullable=True),
-    sa.Column('sizes', sa.JSON(), nullable=True),
-    sa.Column('size_table_photo_path', sa.String(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_photos_id'), 'photos', ['id'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_photos_id'), table_name='photos')
-    op.drop_table('photos')
-    op.drop_index(op.f('ix_orders_id'), table_name='orders')
-    op.drop_table('orders')
-    op.drop_index(op.f('ix_order_items_id'), table_name='order_items')
-    op.drop_table('order_items')
-    # ### end Alembic commands ###
