+++ backend/src/routers/item.py
@@ -14,14 +14,11 @@ from src.path import BASE_DIR
 
 router = APIRouter()
 
-import json
-
 @router.post("/photos", response_model=Photo)
 async def upload_photo(
     name: Optional[str] = Form(None),
     item_description: Optional[str] = Form(None),
     price: float = Form(0.0),
-    sizes: Optional[str] = Form(None),
     file: UploadFile = File(...),
     size_table_photo: Optional[UploadFile] = File(None),
     username: str = Depends(get_current_username),
@@ -38,13 +35,6 @@ async def upload_photo(
         st_filename = await save_file(size_table_photo)
         size_table_photo_path_val = f"/uploads/{st_filename}"
 
-    parsed_sizes = None
-    if sizes:
-        try:
-            parsed_sizes = json.loads(sizes)
-        except json.JSONDecodeError:
-            parsed_sizes = []
-
     new_photo = PhotoDB(
         id=str(uuid.uuid4()),
         filename=file.filename,
@@ -52,7 +42,6 @@ async def upload_photo(
         item_description=item_description,
         path=photo_path,
         price=price,
-        sizes=parsed_sizes,
         size_table_photo_path=size_table_photo_path_val
     )
     db.add(new_photo)
