+++ frontend/admin.html
@@ -30,7 +30,6 @@
                 <input type="text" id="nameInput" name="name" placeholder="Название" data-i18n-key="name">
                 <textarea id="itemDescriptionInput" name="item_description" placeholder="Описание товара" data-i18n-key="item_description"></textarea>
                 <input type="number" id="priceInput" name="price" placeholder="Цена" required min="0" step="0.01" data-i18n-key="price">
-                <input type="text" id="sizesInput" name="sizes" placeholder="Размеры (через запятую)" data-i18n-key="sizes">
                 <button type="submit" data-i18n-key="upload">Загрузить</button>
             </form>
         </section>
