+++ frontend/js/admin.js
@@ -7,7 +7,6 @@ document.addEventListener('DOMContentLoaded', () => {
     const nameInput = document.getElementById('nameInput');
     const itemDescriptionInput = document.getElementById('itemDescriptionInput');
     const priceInput = document.getElementById('priceInput');
-    const sizesInput = document.getElementById('sizesInput');
 
     // --- Функция для отображения галереи ---
     const fetchAndDisplayPhotos = async () => {
@@ -79,7 +78,6 @@ document.addEventListener('DOMContentLoaded', () => {
         const name = nameInput.value;
         const item_description = itemDescriptionInput.value;
         const price = priceInput.value;
-        const sizes = sizesInput.value.split(',').map(s => s.trim()).filter(Boolean);
 
         if (!file || !price) {
             alert('Пожалуйста, выберите файл и укажите цену.');
@@ -94,7 +92,6 @@ document.addEventListener('DOMContentLoaded', () => {
         formData.append('name', name);
         formData.append('item_description', item_description);
         formData.append('price', price);
-        formData.append('sizes', JSON.stringify(sizes));
 
         try {
             const response = await fetch('/photos', {
